requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
aiogram==3.2.0
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
pytoniq==0.1.36
httpx==0.25.2
redis==5.0.1
aioredis==2.0.1
pytonconnect==0.3.2
PyJWT==2.8.0
cryptography==41.0.7
PyNaCl==1.5.0

